+++ UnitySDK/Packages/com.happs.sdk/Runtime/Mobile/Models/MobileSession.cs
@@ -3,6 +3,7 @@ namespace HAppsSDK
 	public sealed class MobileSession
 	{
 		public string DeviceId;
+		public string AppsFlyerKey;
 		public string AccessToken;
 		public long AccessTokenExpiresAtUtc;
 		public string PublicId;